requirements.txt
flask
flask_sqlalchemy
sqlalchemy_jsonapi
flask_apscheduler
blinker
pylxd
